+++ src/checker/providers/classNameProvider.ts
@@ -15,8 +15,28 @@ export class ClassNameDetect implements IDetectProvider {
             return [ConfidenceLevel.Technical, 'javascript'];
         }
 
+        if (/^calc\([^\)]+\)$/m.test(text)) {
+            return [ConfidenceLevel.Technical, 'javascript'];
+        }
+
         if (text.search(/fa[rs]? fa(-[^-]*)*/g) >= 0) {
             return [ConfidenceLevel.Technical, 'font awesome'];
+        }
+
+        if (/^([a-z]+)\s\1-.+$/.test(text)) {
+            // `table table-*`.
+            return [ConfidenceLevel.Technical, 'class'];
+        }
+
+        if (/^query +\{/.test(text)
+            || /^query\(/.test(text)
+            || /^query +[a-zA-Z_][a-zA-Z\d_]+\(/.test(text)
+            || /^mutation\(/.test(text)) {
+            // `query {...}`
+            // `query(...)`
+            // `query myFunction(...)`
+            // `mutation(...)`
+            return [ConfidenceLevel.Technical, 'graphql'];
         }
         
         if (text.search(/btn btn(-[^-]*)*/g) >= 0) {