+++ src/helpers/utils.ts
@@ -1,4 +1,5 @@
 import { ConfidenceLevel } from "../enumerations";
+import { ConfidenceLevelStr } from "../constants";
 
 export const sleep = (ms: number) => new Promise(res => setTimeout(res, ms));
 
@@ -19,13 +20,13 @@ export function distinct<T>(array: Array<T>, compareFn?: (a: T, b: T) => number)
 export function confidenceLevelToString(level: ConfidenceLevel): string {
     switch (level) {
         case ConfidenceLevel.Unknown:
-            return 'Unknown';
+            return ConfidenceLevelStr.Unknown;
 
         case ConfidenceLevel.Technical:
-            return 'Technical';
+            return ConfidenceLevelStr.Technical;
 
         case ConfidenceLevel.Message:
-            return 'Message';
+            return ConfidenceLevelStr.Message;
 
         default:
             throw new Error(`Not supported type ${level}`);
@@ -48,4 +49,36 @@ export function removeSentenceQuotes(text: string | undefined): string | undefin
     }
 
     return text;
+}
+
+export function isCamelCase(text: string): boolean {
+    return /^[a-z][^\s]+$/.test(text);
+}
+
+export function isPascalCase(text: string): boolean {
+    return /^([A-Z][^\sA-Z]+)+$/.test(text);
+}
+
+export function getNonAlphaRatio(text: string | undefined): number {
+    if (!text || text.length <= 0) {
+        return 0;
+    }
+
+    return 1 - stripNonAlphaCharacters(normalizeString(text))!.length / text.length;
+}
+
+export function normalizeString(text: string | undefined): string | undefined {
+    if (!text || text.length <= 0) {
+        return text;
+    }
+
+    return text.normalize('NFD').replace(/[\u0300-\u036f]/g, '');
+}
+
+export function stripNonAlphaCharacters(text: string | undefined): string | undefined {
+    if (!text || text.length <= 0) {
+        return text;
+    }
+
+    return normalizeString(text)!.replace(/[^a-z\d\s]/ig, '');
 }